Deleting a card from list was always deleting the last card
This was an interesting bug where it seemed that whichever
card we click delete button on, the last card was being removed.
But in reality the card entry was correctly being removed from
the list, however details of each card were not being updated
to match the list since we were updating the card contents in the
onInitialized method.
Hence it looked like always the last card is being removed since
the list size got reduced by one on each removal.

The solution was to update the card contents in OnParametersSet
method which is invoked on each list update in the parent
component which resulted in updated parameters being passed to
each child component in the list.