Performance Optimization: TimezoneCard to accept timezone list
Having a list of timezone card components in the Index page was
causing a lot of performance degradation, especially because we
run a timer to continuously refresh all the timezone cards as
they display the current time for their timezone.

Changing from the list of timezone card components to a single
component internally displaying the same list of timezone cards
led to improve performance by a lot (metrics below).

Ideal solution ( next todo item ) is to have only the components
visible at any given time to be updated and not the entire list.

Metrics with the list of components ( 100 timezone cards each in
their own component ):
1.6 second idle out of 21.5 seconds.
        Range: 0 – 21.59 s
        21590 ms
        19191 msScripting
        631 msRendering
        74 msPainting
        75 msSystem
        1618 msIdle
        21590 msTotal

Metrics with a single components displaying the list of 100 cards
internally:
8.5 seconds idle out of 21.1 seconds.
        Range: 0 – 21.19 s
        21193 ms
        11782 msScripting
        575 msRendering
        68 msPainting
        181 msSystem
        8587 msIdle
        21193 msTotal