Adding Google Authentication via Firebase library
Adding reference to BlazorUtils.Firebase library and using it
for Google Authentication.
This library internally uses the Firebase javascript SDK.

We did not need to specify any firebase config for google auth
since this website (timeutilities.web.app) is hosted on firebase.
And this is a convenience provided by firebase, which pulls those
config automatically.
https://firebase.google.com/docs/web/setup#config-object